Enhancement: Turn Example into entity
