+++ src/Entity/Example.php
@@ -1,0 +1,68 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2020 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/symfony-application-template
+ */
+
+namespace Ergebnis\Application\Entity;
+
+use Doctrine\ORM;
+use Ramsey\Uuid;
+
+/**
+ * @ORM\Mapping\Entity
+ * @ORM\Mapping\Table(name="example")
+ */
+class Example
+{
+    /**
+     * @ORM\Mapping\Id
+     * @ORM\Mapping\GeneratedValue(strategy="NONE")
+     * @ORM\Mapping\Column(
+     *     name="id",
+     *     type="string",
+     *     length=36
+     * )
+     *
+     * @var string
+     */
+    private $id;
+
+    /**
+     * @ORM\Mapping\Column(
+     *     name="name",
+     *     type="string"
+     * )
+     *
+     * @var string
+     */
+    private $name;
+
+    private function __construct(string $name)
+    {
+        $this->id = Uuid\Uuid::uuid4()->toString();
+        $this->name = $name;
+    }
+
+    public static function fromName(string $name): self
+    {
+        return new self($name);
+    }
+
+    public function id(): string
+    {
+        return $this->id;
+    }
+
+    public function name(): string
+    {
+        return $this->name;
+    }
+}
